refactor: split management.js and routes/tasks.js into sub-modules (#418)
management.js (1426L) split into:
- management.js (117L) — controls + re-export facade
- management/insights.js (267L) — insight apply/auto-apply/verify
- management/lessons.js (136L) — lesson matching + preflight
- management/evolution.js (236L) — evolution fields, signal trim, edda
- management/dispatch.js (732L) — dispatch plan, model, messages, steps

routes/tasks.js (2177L) split into:
- routes/tasks.js (1103L) — main router + dispatch + CRUD
- routes/tasks-steps.js (292L) — step CRUD/transition/kill/reset/batch
- routes/tasks-lifecycle.js (715L) — cancel/reopen/unblock/rollback/status
- routes/tasks-confidence.js (100L) — confidence/timeline/digest/report

All existing require('./management') callers unchanged (re-export facade).
All API contracts preserved. Integration tests pass.
